refactor(ui): extract shared ToggleView component
Pools, Market, Collectibles, and Gardens each had their own near-identical
ToggleView with the same JSX and only minor styling differences. Replace
them with a single shared component at src/components/ToggleView that
takes optional className (for layout overrides) and id (forwarded to the
underlying IconButtons).

- New shared ToggleView owns the ViewMode enum, eliminating three
  duplicated copies across views/{Collectibles,Market,Greenhouses}/components/types.ts.
- Pools and Collectibles keep their right-margin variant via
  styled(ToggleView) overrides.
- Market and Gardens use the base component directly.
- IDs are forwarded as a prefix (id="clickPool" / id="clickGarden")
  to preserve the existing clickPool*/clickGarden* analytics hooks.
