+++ src/components/ToggleView/index.ts
@@ -1,0 +1,1 @@
+export { default, ViewMode } from './ToggleView'
