--- src/views/Collectibles/components/ToggleView/ToggleView.tsx
@@ -1,38 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-import { ListViewIcon, CardViewIcon, IconButton } from '@plantswap/uikit'
-import { ViewMode } from '../types'
-
-interface ToggleViewProps {
-  viewMode: ViewMode
-  onToggle: (mode: ViewMode) => void
-}
-
-const Container = styled.div`
-  margin-left: -8px;
-
-  ${({ theme }) => theme.mediaQueries.sm} {
-    margin-left: 0;
-  }
-`
-
-const ToggleView: React.FunctionComponent<ToggleViewProps> = ({ viewMode, onToggle }) => {
-  const handleToggle = (mode: ViewMode) => {
-    if (viewMode !== mode) {
-      onToggle(mode)
-    }
-  }
-
-  return (
-    <Container>
-      <IconButton variant="text" scale="sm" id="clickGardenCardView" onClick={() => handleToggle(ViewMode.CARD)}>
-        <CardViewIcon color={viewMode === ViewMode.CARD ? 'primary' : 'textDisabled'} />
-      </IconButton>
-      <IconButton variant="text" scale="sm" id="clickGardenTableView" onClick={() => handleToggle(ViewMode.TABLE)}>
-        <ListViewIcon color={viewMode === ViewMode.TABLE ? 'primary' : 'textDisabled'} />
-      </IconButton>
-    </Container>
-  )
-}
-
-export default ToggleView
