+++ src/views/Collectibles/components/types.ts
@@ -86,8 +86,3 @@ export const DesktopColumnSchema: ColumnsDefTypes[] = [
     label: '',
   },
 ]
-
-export enum ViewMode {
-  'TABLE' = 'TABLE',
-  'CARD' = 'CARD',
-}
